better management of event archiving. fixes issue #33
the cleanup procedure now archive past xref, then checks if there is still 'not archived' xref associated to the event before archiving it too.

A button was added in that backend to do the same (the archive button just put everything in archive, event upcoming upcoming xref associated to the event).